refactor auth pages
